requirements.txt
streamlit
pandas
scikit-learn
joblib
firebase-admin
google-cloud-firestore
google-generativeai